requirements.txt
fastapi>=0.110,<1.0
uvicorn>=0.27,<1.0

pypdf>=4.0,<5.0
python-docx>=1.1,<2.0
openpyxl>=3.1,<4.0
python-pptx>=0.6,<1.0
